refactor(): merge slices into one diagram slice
